requirements.txt
# Core ML/AI Libraries
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.2
langchain>=0.0.220
langchain-community>=0.0.20
faiss-cpu>=1.7.4
chromadb>=0.4.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
python-docx>=0.8.11
PyMuPDF>=1.23.0
pdfplumber>=0.9.0
nltk>=3.8.1

# Language Processing
langdetect>=1.0.9
googletrans==4.0.0rc1
# polyglot>=16.7.4  # Commented out due to build issues on Windows
# fasttext>=0.9.2   # Optional - can be installed separately if needed

# Reinforcement Learning
stable-baselines3>=2.0.0
gymnasium>=0.28.1
tensorboard>=2.13.0

# Web Framework & UI
streamlit>=1.25.0
fastapi>=0.100.0
uvicorn>=0.22.0
gradio>=3.35.0

# Database & Storage
# sqlite3  # Built into Python
redis>=4.5.0
pymongo>=4.4.0

# Communication & Email
# smtplib  # Built into Python
email-validator>=2.0.0
requests>=2.31.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0

# Monitoring & Logging
wandb>=0.15.0
mlflow>=2.4.0
loguru>=0.7.0

# Configuration & Utils
python-dotenv>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0
tqdm>=4.65.0
rich>=13.4.0
typer>=0.9.0

# Local LLM Support
ollama>=0.1.0

# Security
cryptography>=41.0.0